requirements.txt
Django~=5.1
djangorestframework~=3.15.2
celery~=5.4.0
python-dotenv~=1.0.1
requests~=2.32.3
psycopg2==2.9.9
pillow==10.4.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
ipython==8.26.0
coverage==7.6.0
drf-yasg==1.21.7
stripe==10.5.0
redis==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0